test/reftest/vendor/webkit/fast/dom/Window/window-function-name-getter-precedence-expected.txt

PASS: typeof window.print should be function and is.
PASS: typeof window.setTimeout should be function and is.
PASS: typeof window.clearTimeout should be function and is.
PASS: typeof window.setInterval should be function and is.
PASS: typeof window.clearInterval should be function and is.
PASS: typeof window.addEventListener should be function and is.
PASS: typeof window.removeEventListener should be function and is.
PASS: typeof window.eval should be function and is.
PASS: typeof window.parseInt should be function and is.
PASS: typeof window.parseFloat should be function and is.
PASS: typeof window.isNaN should be function and is.
PASS: typeof window.isFinite should be function and is.
PASS: typeof window.escape should be function and is.
PASS: typeof window.unescape should be function and is.
PASS: typeof window.decodeURI should be function and is.
PASS: typeof window.decodeURIComponent should be function and is.
PASS: typeof window.encodeURI should be function and is.
PASS: typeof window.encodeURIComponent should be function and is.
PASS: typeof window.toString should be function and is.
PASS: typeof window.toLocaleString should be function and is.
PASS: typeof window.valueOf should be function and is.
PASS: typeof window.hasOwnProperty should be function and is.
PASS: typeof window.propertyIsEnumerable should be function and is.
PASS: typeof window.isPrototypeOf should be function and is.
PASS: typeof window.myFunction should be function and is.
PASS: typeof window.myPrototypeFunction should be function and is.
